% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package.R
\name{check_desc}
\alias{check_desc}
\title{Check DESCRIPTION file for package fields}
\usage{
check_desc()
}
\value{
logical (TRUE if package fields are present)
}
\description{
Check DESCRIPTION file for package fields
}
\examples{
check_desc()
}
