% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package.R
\name{check_rproj}
\alias{check_rproj}
\title{Check .Rproj file for package fields}
\usage{
check_rproj()
}
\value{
logical (TRUE if package fields are present)
}
\description{
Check .Rproj file for package fields
}
\examples{
check_rproj()
}
