% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_objects.R
\name{gather_all}
\alias{gather_all}
\title{Gather all objects in global environment}
\usage{
gather_all()
}
\value{
all objects in local environment
}
\description{
Gather all objects in global environment
}
\examples{
# f <- function() "top level function"
# gather_all()
}
