% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_classes.R
\name{gather_classes}
\alias{gather_classes}
\title{Gather object classes from global environment}
\usage{
gather_classes()
}
\value{
named vector of class and object
}
\description{
Gather object classes from global environment
}
