% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_objects.R
\name{gather_dfs}
\alias{gather_dfs}
\title{Gather data.frames in global environment}
\usage{
gather_dfs()
}
\value{
character vector of \code{data.frame}s in local environment
}
\description{
Gather data.frames in global environment
}
\examples{
# mtcars <- datasets::mtcars
# attitude <- datasets::attitude
# titanic <- datasets::Titanic
# gather_dfs()
}
