% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_objects.R
\name{gather_lists}
\alias{gather_lists}
\title{Gather lists in global environment}
\usage{
gather_lists()
}
\value{
character vector of \code{list}s in local environment
}
\description{
Gather lists in global environment
}
\examples{
# my_list <- list('mtcars' = datasets::mtcars,
                'attitude' = datasets::attitude,
                'titanic' = datasets::Titanic)
# gather_lists()
}
