% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_objects.R
\name{gather_objects}
\alias{gather_objects}
\title{Gather objects in global environment}
\usage{
gather_objects(type = "all")
}
\value{
all objects in local environment
}
\description{
Gather objects in global environment
}
\examples{
require(R6)
# my_list <- list('mtcars' = datasets::mtcars,
#                 'attitude' = datasets::attitude,
#                'titanic' = datasets::Titanic)
# mtcars <- datasets::mtcars
# class_generator <- R6Class()
# object <- class_generator$new()
# gather_objects(type = "all")
}
