% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_objects.R
\name{gather_r6s}
\alias{gather_r6s}
\title{Gather R6 objects in global environment}
\usage{
gather_r6s()
}
\value{
character vector of \code{R6}s in local environment
}
\description{
Gather R6 objects in global environment
}
\examples{
require(R6)
# class_generator <- R6::R6Class()
# object <- class_generator$new()
# gather_r6s()
}
