% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package.R
\name{is_package}
\alias{is_package}
\title{Check DESCRIPTION and .Rproj files for package fields}
\usage{
is_package()
}
\value{
logical (TRUE if package files/fields are present)
}
\description{
Check DESCRIPTION and .Rproj files for package fields
}
\examples{
is_package()
}
