% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_obj_load.R
\name{mod_obj_load_server}
\alias{mod_obj_load_server}
\title{Load objects module server}
\usage{
mod_obj_load_server(id)
}
\arguments{
\item{id}{module id}
}
\value{
server module function
}
\description{
Load objects module server
}
