% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_obj_load.R
\name{mod_obj_load_ui}
\alias{mod_obj_load_ui}
\title{Load objects module UI}
\usage{
mod_obj_load_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
UI module function
}
\description{
Load objects module UI
}
