% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raiden-package.R
\docType{package}
\name{raiden-package}
\alias{raiden}
\alias{raiden-package}
\title{raiden: roxygen2 addin for documentation in R}
\description{
Easily document your R functions, objects, and methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mjfrigaard.github.io/raiden/}
}

}
\keyword{internal}
