% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raiden_app.R
\name{raiden_app}
\alias{raiden_app}
\title{Launch raiden addin}
\usage{
raiden_app()
}
\description{
Launch the \code{raidenApp()} Addin to document your R object. For more
information on \code{.rs.invokeShinyPaneViewer}, see
\href{https://bit.ly/44QNM1z}{this SO post}
}
