% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raidenApp.R
\name{raidenApp}
\alias{raidenApp}
\title{Launch raiden addin}
\usage{
raidenApp()
}
\description{
Launch the \code{raidenApp()} Addin to document your R object. For more
information on \code{.rs.invokeShinyPaneViewer}, see
\href{https://stackoverflow.com/questions/35311318/opening-shiny-app-directly-in-the-default-browser}{this SO post}
}
