% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_classes.R
\name{squish_class}
\alias{squish_class}
\title{Squish \code{class()}}
\usage{
squish_class(x)
}
\arguments{
\item{x}{object}
}
\value{
class as \code{length(x) == 1}
}
\description{
Squish \code{class()}
}
\examples{
squish_class(Sys.time())
}
